# Speed up arbtt-stats --filter