@nomeata 2021-02-21
>  I'll need to deal with log rotation sooner or later, but this fruit was hanging lower. :-))

I haven’t rotated my log ever, and I have been using it since 2009 :-) (But then, I am a hoarder: I collect the data, but rarely look at it.)

I would prefer it if timezone handling wouldn’t be shoved under the rug, though – my stats have been recorded in various continents, and if I want to know what I tend to do after midnight, that certainly should be local time.

We could maybe detect when the filter does not mention `$time`.

Or maybe the problem (with the original code) is that although `applyCond` was set up to be partially applied (i.e. parse the condition once, and use for many `TimeLogEntry`), this doesn’t work because the Timezone comes earlier, and so in `filterPredicate` there is no gain.

So the solution might be that parsing the condition does not depend on `TimeZone`, and the output of the parser (a `ApplyCond` of sorts) takes the timezone of the sample. 

Let me give it a shot…