@liskin 2021-02-21
I should have been more clear that this only speeds things up, it doesn't change behaviour. The code never considered the timezone of the sample, because the timezone of samples was never recorded.