@liskin 2021-02-21
> I haven’t rotated my log ever, and I have been using it since 2009 :-) (But then, I am a hoarder: I collect the data, but rarely look at it.)

Oh, how long does your arbtt-stats take then? I have `--sample-rate=20` and logs going back to December 2020, which I presume is about the same as the default sample rate and logs going back to cca April/May 2020, so logs going back 10 more years should take more than 10 seconds to analyze, unless something else is seriously wrong here.