@nomeata 2021-02-21
Yes, takes more than 10s :)

I use the default sample rate of 60s, and I get:
```
~ $ time arbtt-stats --filter="\$sampleage < 24:00"
Total time per tag                                                                                                                                                                                  
…
real	0m58,102s
user	0m57,731s
sys	0m0,364s
```
I have 1773305 samples, 
```
~/projekte/programming/arbtt $ ls -sh ~/.arbtt/capture.log
180M /home/jojo/.arbtt/capture.log
```